requirements.txt
langchain>=0.1.0
python-dotenv>=1.0.0
openai>=1.12.0
pypdf>=4.0.0
python-docx>=1.0.0
nbformat>=5.9.0
streamlit>=1.32.0 